requirements.txt
colorama==0.4.6
iniconfig==2.1.0
mysql-connector-python==9.4.0
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
